# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MindSafe Mental Health Platform API
  version: 1.0.0
  description: |
    Privacy-first AI mental health platform. All endpoints under `/api` are served by the
    API Gateway (Express) which proxies ML operations to internal Python microservices.

    ## Authentication
    Most endpoints require a Bearer token in the `Authorization` header.
    Obtain tokens via `/api/login` or `/api/auth/google`. Access tokens expire in 15 minutes;
    use `/api/refresh-token` to get a new pair.

    ## Error Format
    All errors return `{ "message": "..." }` with an appropriate HTTP status code.
  contact:
    name: MindSafe Team
  license:
    name: Private
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://mindsafe-api.onrender.com
    description: Free-tier production (Render)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/api/login`. Expires in 15 minutes.
  schemas:
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "Invalid or expired token"
...
    MoodLogRequest:
      type: object
      required: [mood_score]
      properties:
        mood_score:
          type: integer
          minimum: 1
          maximum: 10
          example: 7
        mood_label:
          type: string
          example: happy
        notes:
          type: string
          nullable: true
          example: "Had a good therapy session today"
        emotion_scores:
          type: object
          nullable: true
          additionalProperties:
            type: number
          example: { "joy": 0.8, "sadness": 0.1 }
        logged_at:
          type: string
          format: date-time
          nullable: true
    MoodLogResponse:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: string
        logged_date:
          type: string
          format: date
        mood_score:
          type: integer
        mood_label:
          type: string
        notes:
          type: string
        emotion_scores:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /api/moods/log:
    post:
      tags: [Mood Tracking]
      summary: Log today's mood
      operationId: logMood
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoodLogRequest"
            example:
              mood_score: 7
              mood_label: happy
              notes: "Had a good therapy session today"
              emotion_scores:
                joy: 0.8
                sadness: 0.1
      responses:
        "201":
          description: Mood logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoodLogResponse"
        "400":
          description: mood_score must be between 1 and 10
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: Mood service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...